update appbar widget
